requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
oauthlib==2.0.6
py-trello==0.10.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
wsgiref==0.1.2
